Refactor error handling in parser and state modules; update version to 0.3.0
- Enhanced error reporting in `parse_hand_internal`, `parse_tile`, and `parse_meld` functions to include input context.
- Changed error handling in `GameState` to use `InvalidState` for replay desynchronization errors.
- Updated package version in `Cargo.toml` and `uv.lock` to 0.3.0.
