+++ riichienv-core/src/hand_evaluator.rs
@@ -253,6 +253,29 @@ impl HandEvaluator {
     }
 }
 
+pub fn check_riichi_candidates(tiles_136: Vec<u8>) -> Vec<u32> {
+    let mut candidates = Vec::new();
+    // Convert to 34-tile hand
+    let mut tiles_34 = Vec::with_capacity(tiles_136.len());
+    for t in &tiles_136 {
+        tiles_34.push(t / 4);
+    }
+
+    for (i, &t_discard) in tiles_136.iter().enumerate() {
+        let mut hand = crate::types::Hand::default();
+        for (j, &t) in tiles_34.iter().enumerate() {
+            if i != j {
+                hand.add(t);
+            }
+        }
+
+        if agari::is_tenpai(&mut hand) {
+            candidates.push(t_discard as u32);
+        }
+    }
+    candidates
+}
+
 fn get_next_tile(tile: u8) -> u8 {
     if tile < 9 {
         if tile == 8 {
