+++ riichienv-python/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "riichienv-python"
-version = "0.2.1"
+version = "0.3.0"
 edition = "2021"
 license = "Apache-2.0"
